refactor(domain): move TocEntry to core/domain with @Serializable
Move TocEntry from app/feature/reader to core/domain and add @Serializable
annotation to enable JSON caching in Room for downstream tasks. Update all
imports in TocParser and EpubReaderViewModel accordingly.
