+++ core/domain/src/main/kotlin/com/riffle/core/domain/TocEntry.kt
@@ -1,5 +1,8 @@
-package com.riffle.app.feature.reader
+package com.riffle.core.domain
 
+import kotlinx.serialization.Serializable
+
+@Serializable
 data class TocEntry(
     val title: String,
     val href: String,
